+++ src/components/AutoComplete/AutoCompleteSelected.js
@@ -2,7 +2,7 @@ import React, { useRef, useCallback } from 'react'
 import PropTypes from 'prop-types'
 import { identity, noop } from '../../utils'
 import { theme } from '../../theme-legacy'
-import { ButtonBase } from '../Button/ButtonBase'
+import ButtonBase from '../ButtonBase/ButtonBase'
 import AutoComplete from './AutoComplete'
 
 function AutoCompleteSelected({
