+++ src/components/Button/ButtonText.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { ButtonBase } from './ButtonBase'
+import ButtonBase from '../ButtonBase/ButtonBase'
 import { GU } from '../../style'
 import { useTheme } from '../../theme'
