+++ src/components/DateRange/DatePicker.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled, { css } from 'styled-components'
 import dayjs from 'dayjs'
-import { ButtonBase } from '../Button/ButtonBase'
+import ButtonBase from '../ButtonBase/ButtonBase'
 import Text from '../Text/Text'
 import { GU } from '../../style'
 import { useTheme } from '../../theme'
